fix URL to point to correct file on Figshare
Open Refine lesson uses a special file version of the ecology dataset created to show features of Open Refine. URL in this lesson about scripts was pointing to the messy_data.xls which is for the better spreadsheet practices lesson. I updated the URL to point to the Portal_rodents file on figshare.