Update 01-working-with-openrefine.md
OpenRefine now defaults to trimming leading/trailing whitespace upon import. In the Split section #6 and the solution will need to be modified. I just ran through and there was no cols called scientificName 3 and scientificName 4 because of the trimming. In addition the Trim Leading ... section exercise needs to be modified because the results will be 0 cells modified. 
I only added a sentence to Trim because I'm not sure how to handle this. It requires some major work.