Remove reference to sorting column values as numbers
The "exploring numeric columns" section contained the following sentence: "We saw earlier how we can sort column values as numbers, but this does not change the cells in a column from text to numbers. Rather, this interprets the values as numbers for the purposes of sorting but keeps the underlying data type as is"

This episode has covered sorting facet values, but not column values, so this sentence does not make sense here and should be removed to avoid learner confusion.