Cannot cluster as instructed because sciNames gone
This episode. Oof.
This episode already split scientificNames into two separate columns after trimming white space.  Are we supposed to remember to have unchecked the box to delete the original column in the first set of instructions?  I can go add that command instead.
Learners cannot cluster on two columns.  Or should the instructions focus on just species?  That would be a few replacements of scientificName with species and a few edits to the demo text.  
Suggestions welcome.